chore: relocate components/preview files to their respective new locations
